FTVPI-31: Fixed incorrect current time calculation
[Problem]
All video frames are dropped at the beginning of playback

[Root Cause]
AudioTrack module is the provider of the media clock for A/V synchronization.
A typo error caused error in calculating startMediaTimeUs that caused error in
calculating the current time. This resulted into video frames being dropped till
it caught up with the incorrect current time. startMediaTimeState variable
is used to track the state of the startMediaTimeUs variable. However, in
two places, startMediaTimeUs  was being used instead of startMediaTimeState.
This looks like a silly typo error

[Solution]
Correct the typo errors by using startMediaTimeState correctly.

Change-Id: I14d2238cc4fada1ef63f29121574270e1847bbc2
Reviewed-on: http://e-gerrit.labcollab.net/28202
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-on: http://e-gerrit.labcollab.net/29981