FTVPI-32: Improved video size detection and notification.
[Problem]
1. The first onVideoSizeChanged callback gives -1 as width and height.

2. The onVideoSizeChanged callback with valid size values comes after
state changed to READY.The app has no way to find out the right
video size by the time prepare phase is completed.

[Root Cause]
1. The PixelWidthHeightRatio is detected when media codec returns
INFO_INPUT_FORMAT_CHANGED. Since the lastReportedPIxelWidthHeightRatio and
currentPixelWidthHeightRatio changes in maybeNotifyVideoSizeChanged,
we notify the app with correct PixelWidthHeightRatio but invalid video size -1.

2. Output size is detected when the mediacodec returns INFO_OUTPUT_FORMAT_CHANGED.
However, it is notified to the app via onVideoSizeCallback only when
the first frame is rendered. If the initial video frames are dropped due to
A/V sync logic, it delays notifying the app about video size.

[Solution]
1. To prevent invalid width and height notification, added a check not to notify
if the video size is not yet detected.

2. We don't have to wait till we render first frame to notify the video size to app.
Notify it as soon as we detect a valid size.

Change-Id: Ie217f6040ccd7868f6da6c0f56ab9e2a719f58c8
Reviewed-on: http://e-gerrit.labcollab.net/28203
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-on: http://e-gerrit.labcollab.net/29983