FTVPI-34: Fix for Codec error VIDC_1080P_ERROR_HEADER_NOT_FOUND
[Problem]
Decoder throws error VIDC_1080P_ERROR_HEADER_NOT_FOUND on decoding frames
initially.

[Root Cause]
In smoothstreaming, some content does not contain config data
appended as part of first frame. Also, Config data is not appended
to the first frame in feedInputBuffer. This happens because when
the codec is created, the codecReconfigurationState is not set to
RECONFIGURATION_STATE_WRITE_PENDING.

[Solution]
When we create the codec, we detect if the codec supports reconfiguration
and we set the codecReconfigurationState to RECONFIGURATION_STATE_WRITE_PENDING.
Note that for audio, we don't need to do this, and hence we depend on canReconfigureCodec
to decide if we need to append config data to first frame or not.

[Tests]
Tested on Fire TV, stick and tablets.

Change-Id: I1246bdae5d56554e634258466f1d449c5694d641
Reviewed-on: http://e-gerrit.labcollab.net/28205
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>