FTVPI-22: Fix for jitter on resolution switch
[Problem]
In HLS, Smoothstreaming and DASH, when the resolution switches,
we see a jitter in video.

[Root Cause]
Video decoders on FireTV family supports adaptive playback by default
This means we can reconfigure the decoder to start decoding new resolution without
recreating it. In API Level 19 and above, CodecCapabilities has added
an API which can be queried to find out if the codec is adaptive or not.
Even though some implementation of this API was pulled into FireTV, it is
not complete - i.e low level OMX codec does not indicate that it supports it.

[Solution]
Video decoders in all amazon devices by default support adaptive streaming capability.
We add a quirk to identify that video decoder supports adaptive for all amazon
tablets.

[Limitation]
Currently, we have enabled this feature only for MPEG4 and H.264. For
other video codecs we need to do it based on need in future.

[Tests]
Tested on Fire TV , stick and tablets.

Change-Id: I040c4e6ba127aeaf35f32425c2c99a7772f58000
Reviewed-on: http://e-gerrit.labcollab.net/28579
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-on: https://e-gerrit.labcollab.net/33734
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Tested-by: Zsolt Matyas <zmmatyas@amazon.com>