FTVPI-4: Fix for handling audio latencies inconsistency in FireOS4 devices
[Problem]
1. Video playback starts, but stops immediately on FireOS 4 tablets.
2. Sometimes on seeking, playback stops on FireOS 4 tablets.

[Root cause]
In QCOM and MTK based Amazon devices running KitKat,
the AudioTrack (specifically getPlaybackHeadPosition implementation) has several issues.

AudioTrack accounts for audio latencies in the head position returned by
getPlaybackHeadPosition API.
Due to this, during the latency  stablization phase that happens at start of audio
playbak, it no longer returns a continuously advancing frame position counter, instead
can dip a bit till steady state is reached.
This happens at the start of playback as well as after every seek.

Additionally,this API pre-deducts software latencies from the head position.
Exoplayer is dependent on both getTimeStamp and getPlaybackHeadPosition APIs and
it does not expect latencies to be pre-deducted from the returned value.
Instead, ExoPlayer calls getLatency java API of AudioTrack to compute hardware latencies
and deducts it from the current audio position.

MediaController sometimes issues duplicate seek calls to the media player
on touching the progress bar once.
On a seek operation, ExoPlayer pauses the audio track.
When such duplicate quick seeks are issued to exoPlayer, the sequence of calls
to audio track are pause, play and pause in quick succession.
In such a scenario Audio track does not pause rendering audio immediately
and continues to increment the playback head position.

But since pause was issued immediately after a play, the audio track has not yet
stablized the latency calculation and after few iterations we see a dip
in the playback head position.

Moreover, sometimes, even after adding latency compenstation during the stablization
phase, the compensated playback head position is a negative number - in terms of playback
head position a very large number because the returned value even though it is an int
is considered as an unsigned integer value as per Android documentation.
http://developer.android.com/reference/android/media/AudioTrack.html#getPlaybackHeadPosition()

These issues results in large jump in calculated play head position and audio current position.

[Solution]
Add back the software latencies to getPlaybackHeadPosition returned values in AudioTrack
module of ExoPlayer when either in playing state or when in paused state and current play
head position valie is non-zero.

Improve playback head position wrap around detection logic to handle the dips.

Set the HW playback latencies to be around 90 ms that matches the latencies
for all amazon based tablet devices.

If within 1 second of start of playback, the play head position is negative,
reset it to 0.

[Tests] Tested video playback, seek and audio/video sync in
Fire Tablets & phone.

Change-Id: Iba6d352594053eeb850dd52e895deba8ffdfc100
Reviewed-on: https://e-gerrit.labcollab.net/33735
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Tested-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-on: https://e-gerrit.labcollab.net/34087