FTVPI-4: Fix for handling audio latencies in FireOS4 devices - part 2
[Problem]
Sometimes on seeking, playback stops on FireOS 4 tablets. We see that
the calculated timestamps jump to a very large value. This is not seen after
every seek.

[Root cause]
When a seek results into an immediate buffering, audio track,
in such a scenario, does not pause rendering audio immediately
and continues to increment the playback head position.

But since pause was issued immediately after a play,
the audio track has not yet stablized the latency calculation and after few
iterations we see a dip in the playback head position.

The dependent patch https://e-gerrit.labcollab.net/#/c/28462/
compensates for this dip only if the track is known to be in playing state.

Hence, this results into an incorrect timestamp calculation.

[Solution]
We now compensate for the latency even in paused sate, but if and only if the
playback head position is non-zero. This is done, because otherwise,
when the playback starts initially, the playback head position
is always zero and adding a latency value will make hasPendingData to misbehave
and it always return false. Hence the playback will never start.
Since we know that pause is triggered only after few hundered milliseconds of play,
the play head position will never be zero and it is safe to depend on play head position
value.

[Tests] Tested video playback, seek and audio/video sync in
Fire Tablets & phone.

Change-Id: I7e59e25f8b557b86331f7efa85dc1226a295b29d
Reviewed-on: http://e-gerrit.labcollab.net/29061
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>