FTVPI-106: Fix dolby passthrough seek issues
[Problem]
After seek, video takes lot of time to start playback while
audio start immediately.

[Root Cause]
getTimestamp API in FOS5 may return false (unlike in 1.5x where
it always returns true) just after creating the AudioTrack or after reset.
The returned value of getTimestamp was not being considered
for timestamp calculation which resulted into incorrect timestamp.
This cause A/V sync logic to kick in to delay starting of video.

[Solution] Check for returned boolean value by getTimestamp API before
using the timestamp. If it returns false, we just use the startMediaTimeUs

[Tests] Tested on Fire TV FOS5 builds. Fire TV stick has some other platform
issues in FOS5 because of which we could not validate

Change-Id: Ica81452a489610b7198a5047f51bbc00116e8fba
Reviewed-on: https://e-gerrit.labcollab.net/32674
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-on: https://e-gerrit.labcollab.net/33739
Tested-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-on: https://e-gerrit.labcollab.net/34091