FTVPI-159: Threading issue- State change callbacks on incorrect threads [1/2]
This patch introduces one new constructor in the factory to allow
sending ExoPlayer a looper to be used for event callbacks
from ExoPlayer to whoever is using it.

This is in keeping with the way other components also provide a
choice of the event handler or event looper to be used
for event callbacks.

This change will be used by AMZNMediaPlayer to request exoPlayer
events on the state machine's looper thereby eliminating the
threading issue identified.

Change-Id: Iec4c346d14a2780462589fd28bead97f9d3d1c92
Reviewed-on: https://e-gerrit.labcollab.net/32851
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Tested-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-on: https://e-gerrit.labcollab.net/33743
Reviewed-by: Kexin Zhang <kexinz@amazon.com>
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Tested-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-on: https://e-gerrit.labcollab.net/34093
Reviewed-on: https://e-gerrit.labcollab.net/36072