FTVPI-200: onLoad* callbacks didn't have bitrate information for HLS when type = TYPE_MANIFEST.
[Problem]
Exoplayer code wasn't currently populating the
format field for MediaPlaylistChunk which is used when doing chunk
operation for Manifest.

[Solution]
This patch adds the format object (if available) to the MediaPlayerlistChunk
when chunk operation is done in getChunkOperation. The format object
is obtained from the variants using the supplied variantIndex.

The callbacks for bitrate are now present and match for type_media and type_manifest.

Change-Id: Idb3b9087916e1cefe38391cb2cafa05d63cd134d
Reviewed-on: https://e-gerrit.labcollab.net/33858
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Tested-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-on: https://e-gerrit.labcollab.net/34032
Reviewed-on: https://e-gerrit.labcollab.net/34095
Reviewed-by: Gaurav Gangwal <gangwal@lab126.com>
Tested-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-on: https://e-gerrit.labcollab.net/36074
Reviewed-on: https://e-gerrit.labcollab.net/36689
Reviewed-on: https://e-gerrit.labcollab.net/37759
Reviewed-on: https://e-gerrit.labcollab.net/38319