FTVVIZ-626: MP3 playback may terminate prematurely when playing it on FireTV Gen2 device.
[Problem]
Exoplayer was defaulting to using MTK's mp3 decoder in FireTV Gen2.
The MTK decoder has an issue that it incorrectly passes the timestamp as part
of output buffer leading to a trigger of "discontinuity detection" logic inside
ExoPlayer which results in a pre-mature end of playback.

[Solution]
This patch works around the issue by filtering out MTK's mp3
decoder on FireTV gen2. This leads to the default google one being used.

Change-Id: I8d69540f2f725e2780611ea72a14d5d8530fc4a8
Reviewed-on: https://e-gerrit.labcollab.net/34730
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Tested-by: Tarun Solanki <tarunsol@lab126.com>