FTVPI-294:  Exo player stuck in prepare state forever
[Patch for 1.3.3 only. Not sure if 1.4.2 has this issue
as well]

[Problem]
Vizzini HLS chunk playback would sometimes get stuck in prepare
and never come out. After much debugging the problem was
traced down to the HlsSampleSource and was happening
due to a race condition.

Basically, the prepare is called in a polling loop every 10msec.
The chunk download proceeds indepedently, and the data is read
by the TSExtractor. Once it has enough data, it does some processing
in it's consume function and sets up a flag. This flag is then read
by HlsSampleSource/HlsExtractorWrapper on the prepare call
and everything is great.

However, if the chunkOnLoadComplete happens to preceede the prepare
call, because it doesn't see any tracks, it proceeds to clear the
extractor state which also clears the previous flag. Therefore
on the next call of prepare (which is run on same thread), it fails
to find anything and simply keeps trying.

It seems the Exoplayer team wrote this with the assumption that the
TS chunk would be sufficiently big enough to never run into a
situation where the chunk load finishes before the TS extractor had
time to set the flag and prepare to process it.

[Solution]
Added a check to not clear state or trim allocator is chunk type is
TsChunk and the sample source isn't preapared in onLoadCompleted
of HlsSampleSource.

Ideally we shouldn't see this in anyone else's case if their
TS chunk is big enough.

We also still need to understand if this adversely affect the case
where all tracks are disabled but sample source is still running
(shouldn't happen in real world scenario).

[Testing]
Verified by zmmatyas and was able to freely play the hls content
without reproducing the issue.

Change-Id: Iee653bc90ce4c9c137cb819defbd169f4e52df49
Reviewed-on: https://e-gerrit.labcollab.net/35428
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Tested-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-by: Gaurav Gangwal <gangwal@lab126.com>
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>