FTVPI-464: HBOGo doesn't play content after upgrading to ExoPlayer 1.4.2
Extract PlayReady header from PSSH for MediaDrm

[Problem]
In ExoPlayer 1.4.1, a change was introduced that sends the entire PSSH box to
MediaDrm instead of just the PlayReady header.  This was done to fix an issue
with SmoothStreaming on Android TV.

However, PlayReady support on Amazon devices currently expects the PlayReady
header only -- not the entire PSSH box.  Modifying Amazon devices to handle the
entire PSSH box would require changes to three different PlayReady
implementations, so for now it is better to simply restore ExoPlayer behavior
to pre-1.4.1 behavior.

[Solution]
For Amazon devices, extract the PlayReady header from the PSSH box prior to
sending it to MediaDrm.

[Test]
- Verified playback works with HBOGo content shell on Bueller 1.6.3 and Montoya
  1.2.1.
- Verified MediaDrm.getKeyRequest() succeeds on FireOS 5 with Montoya, Bueller,
  and Sloane (other codec initialization issues still exist, and will be
  addressed later).

Change-Id: I21b87b1f68970ff432bd6ada17d263f4498a5a12
Reviewed-on: https://e-gerrit.labcollab.net/36530
Tested-by: Don Darling <darling@lab126.com>
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-by: Don Darling <darling@lab126.com>