FTVPI-464: Extract PlayReady header from PSSH for MediaDrm on FireTV devices.
Extract PlayReady header from PSSH for MediaDrm

[Problem]
In ExoPlayer 1.4.1, a change was introduced that sends the entire PSSH box
to MediaDrm instead of just the PlayReady header.  This was done to fix an
issue with SmoothStreaming on Android TV.

However, PlayReady support on Amazon Fire TV devices currently expects the
PlayReady header only -- not the entire PSSH box.  Modifying Amazon devices
to handle the entire PSSH box would require changes to three different
PlayReady implementations, so for now it is better to simply restore
ExoPlayer behavior to pre-1.4.1 behavior.

[Solution]
For Amazon devices, extract the PlayReady header from the PSSH box prior to
sending it to MediaDrm.

[Test]
- Verified playback works with Playready smoothstreaming content on Bueller
  1.6.3 and Montoya 1.2.1.
- Verified MediaDrm.getKeyRequest() succeeds on FireOS 5 with Montoya,
  Bueller, and Sloane (other codec initialization issues still exist, and
  will be addressed later).

Change-Id: I5b22087ba8d3c687be85acea851d4c976ab6f5b8
Reviewed-on: https://e-gerrit.labcollab.net/47979
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>