FTVPI-568: FTVPI-800: Workaround to avoid sending CSD data in Configure API of Video decoder Fix for exception at the end of Playready content playback in Fire TV Gen2
[Problem]
- Video decoder configuration failed in playback of Playready encrypted content
- When the PlayReady encrypted playback ends and the player sends an empty frame to MediaCodec
with  BUFFER_FLAG_END_OF_STREAM flag set, MediaCodec throws an exception.

[Root cause]
- FireTV Gen2 has a platform limitation that it cannot handle clear data sent
in secure buffer and this happens when the application sends CSD data in clear
as part of MediaFormat  and also provides a valid non-null
crypto object in configure API.
- FireTV Gen2 has a platform limitation that it cannot handle empty clear buffer in
queueInputBuffer with BUFFER_FLAG_END_OF_STREAM flag set if the MediaCodec is
associated with a valid Crypto object.

[Woraround]
Till this issue is fixed in platform, we need to avoid sending CSD
data as part of MediaFormat in configure API for video decoder only.
Instead this should be sent in clear as part of first frame which is already
being done as part of "FTVPI-34: Fix for Codec error VIDC_1080P_ERROR_HEADER_NOT_FOUND"

Platform fix is internally tracked by SLAMFOSV-5597

[Target Devices]
Fire TV Gen2 only.

[Solution]
Since this issue is fixed in future OS version as per SLAMFOSV-5597,
enable the codecNeedsEosPropagationWorkaround workaround for Fire TV Gen2 for
the OS versions with this issue. This workaround avoids sending empty frame
with BUFFER_FLAG_END_OF_STREAM flag, and handles the end of stream detection
in the player differently.

[Tests]
Tested encrypted and clear content playback on Fire TV Gen2.

[Notes] This workaround will work even after the platform fix.
This workaround applies only for Fire TV Gen2

Change-Id: Ib90cfa1f620722ea8d153bd2e352af47235c0a0f
Reviewed-on: https://e-gerrit.labcollab.net/40665
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Tested-by: Zsolt Matyas <zmmatyas@amazon.com>