FTVPI-632: Workaround to handle platform limitation of max input size for AVC for Fire TV Gen2 only
[Problem]
Certain 4K videos crash on playing.

[Root cause]
Fire TV Gen2 device has a platform limitation that it does not support
max input size configuration more than 2.8 MB for AVC decoder. Exoplayer
calculates the max input size based on the resolution and tries to
configure the AVC decoder with that value which in case of a 4K video
exceeds the platform limit of 2.8 MB.

[Solution]
We avoid setting max input size more than 2.8 MB for AVC decoder on FireTV Gen2 device.
Instead, use the defaults.

[Tests]
Tested the specific 4K content playback on Fire TV Gen2 that used to fail earlier
Tested other content for regressions.

[Notes]
This workaround is applicable only for the use case where the sample source
has not configured the max input size as required by the content.
It is possible that certain 4K videos that really need more than 2.8MB per frame
still fail to play on Fire TV Gen2 even if this workaround is triggered.

Platform team is working on removing this limitation for non-secure content
only. This is internally tracked by SLAMFOSV-5878. We will modify this
patch to affect only secure content after this is fixed.

Change-Id: Iddff8b1c316644925e67e0ac0f16c4435609a158
Reviewed-on: https://e-gerrit.labcollab.net/37224
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-on: https://e-gerrit.labcollab.net/37769
Tested-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-on: https://e-gerrit.labcollab.net/38326