FTVPI-772: Fix for frame drops during dolby pass-through playback in Fire TV Gen2
[Problem]
During playback of Dolby content in dolby pass-through mode, the player drops too many frames.

[Root cause]
Version 1.5.3 introduced a new feature of reporting audio track under-run to the application.
The player calls getPlaybackHeadPosition of AudioTrack to decide if the audio track has
under-run or not. Due to the design limitation in the platform, this API sometimes takes upto 20 msec
to return, thereby delaying the processing of video pipeline which results into frequent video frame drops.

[Solution]
We avoid reporting audio track under-run for Fire TV Gen2

[Tests]
Tested the Dolby pass-through playback on Fire TV Gen2.

[Notes]
This patch removes the feature of reporting audio track under-runs to the app.

Change-Id: Ic4529bf25edda45641891363e4969cfc723c9951
Reviewed-on: https://e-gerrit.labcollab.net/37777
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>