FTVPI-800: Fix for exception at the end of Playready content playback in Fire TV Gen2
[Problem]
When the PlayReady encrypted playback ends and the player sends an empty frame to MediaCodec
with  BUFFER_FLAG_END_OF_STREAM flag set, MediaCodec throws an exception.

[Root cause]
FireTV Gen2 has a platform limitation that it cannot handle empty clear buffer in
queueInputBuffer with BUFFER_FLAG_END_OF_STREAM flag set if the MediaCodec is
associated with a valid Crypto object.

[Solution]
Since this issue is fixed in future OS version as per SLAMFOSV-5597,
enable the codecNeedsEosPropagationWorkaround workaround for Fire TV Gen2 for
the OS versions with this issue. This workaround avoids sending empty frame
with BUFFER_FLAG_END_OF_STREAM flag, and handles the end of stream detection
in the player differently.

Change-Id: I57ba045553549e177da71e90204d93ca52d69250
Reviewed-on: https://e-gerrit.labcollab.net/38088
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>