FTVPI-1555: [EXO2] Playready header issues
Original issue on ExoPlayer v1:
FTVPI-464: Extract PlayReady header from PSSH for MediaDrm on FireTV devices.

Extract PlayReady header from PSSH for MediaDrm

[Problem]
In ExoPlayer 1.4.1, a change was introduced that sends the entire PSSH box
to MediaDrm instead of just the PlayReady header.  This was done to fix an
issue with SmoothStreaming on Android TV.

However, PlayReady support on Amazon Fire TV devices currently expects the
PlayReady header only -- not the entire PSSH box.  Modifying Amazon devices
to handle the entire PSSH box would require changes to three different
PlayReady implementations, so for now it is better to simply restore
ExoPlayer behavior to pre-1.4.1 behavior.

[Solution]
For Amazon devices, extract the PlayReady header from the PSSH box prior to
sending it to MediaDrm.

[Test]
- Verified playback works with Playready smoothstreaming content on Bueller
  1.6.3 and Montoya 1.2.1.
- Verified MediaDrm.getKeyRequest() succeeds on FireOS 5 with Montoya,
  Bueller, and Sloane (other codec initialization issues still exist, and
  will be addressed later).

Change-Id: Ida2187e2bc960d74733e7af5e4f87cb9ad9017e3
Reviewed-on: https://e-gerrit.labcollab.net/47028
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>
Reviewed-on: https://e-gerrit.labcollab.net/48448
Reviewed-by: Tarun Solanki <tarunsol@lab126.com>
Reviewed-on: https://e-gerrit.labcollab.net/49057
Reviewed-by: Zsolt Matyas <zmmatyas@amazon.com>
Reviewed-on: https://e-gerrit.labcollab.net/49318
Reviewed-on: https://e-gerrit.labcollab.net/50699
Reviewed-on: https://e-gerrit.labcollab.net/51280