FTVPI-2015: Fix for handling discontinuities in EAC3 and AC3
[Problem]
Workaround to support Dolby passthrough in FTV devices has a regression
that it does not support PTS discontinuities.

[Root cause]
When the original patch to work around Dolby passthrough was done, Exo did
not handle the discontinuities in Dolby audio at all. In later versions
this was fixed. The workaround needs to be slightly modified to leverage that.

[Solution]
Remove the workaround done for setting the start media time and enable
the path to handle discontinuities

[Tests]
Tested Dolby passthrough  playback on FTV Box Gen1

Change-Id: I8de6e1dcfcda8fa161b0ac40d0839d1500d42239
Reviewed-on: https://e-gerrit.labcollab.net/47074
Reviewed-by: Srikanth Peddibhotla <peddisri@lab126.com>
Tested-by: Srikanth Peddibhotla <peddisri@lab126.com>