FTVPI-2218: Support for disabling snapping video frame release times to VSYNC added
[Problem]
24 FPS 4K content when played on Fire TV Box Gen2 with a mode switch to 4K@24Hz
has judders in video.

[Root cause]
Due to snapping to VSYNC, VideoFrameReleaseTimeHelper generates video frames with
timestamp gap pattern of 50/33/41 msec. Ideally, it should be spaced 41.6 msec apart.
Till we  find the root cause, apps need to be able to disable snapping to VSYNC.

[Solution]
Added an API in AmazonQuirks to disable snapping to vsync. Used this boolean to
control the behaviour of VideoFrameReleaseTimeHelper