FTVPI -2396: Blanche switch to BT while playing Dolby doesn't work
[Problem] Switching audio from dolby out to bt headset causes playback to stop when using the demo app. In FOS5 the platform would send an intent HDMI_AUDIO_PLUG when audio was switched from dolby out to BT headset. This does not happen in FOS6 devices like Blanche.

[Solution] Add new logic for Blanceh+ use cases. Platform will expose a new flag 'use_external_surround_sound_flag' on Blanche and newer Fire Televisions. If this flag is set then Exo player should only depend on the setting 'external_surround_sound_enabled' to determine whether an audio output supports dolby.

If the 'use_external_surround_sound_flag' flag is not set. Then Exo player will internally also register for HDMI_AUDIO_PLUG intent. When the 'external_surround_sound_enabled' is set it indicates that device is connected to a dolby capable output. But when 'external_surround_sound_enabled' is not set, then Exoplayer fallsback to the audio capabilities from the most recent HDMI_AUDIO_PLUG intent.

Add the onAudioCapabilities callback to the demo app. In this callback the demo app rebuilds the the media pipeline when there is a change in audio output deivce.

[Devices]
FOS6 fire tv

cr https://code.amazon.com/reviews/CR-3156827